20251112_2341

+++ src/main/java/capstone/user/repository/UserRepository.java
@@ -2,10 +2,9 @@ package capstone.user.repository;
 
 import capstone.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     boolean existsByEmail(String email);
     Optional<User> findByEmail(String email);
 }
